proxy: lua API version 2
See t/proxyfuncgen.lua for documentation on API changes.

See also https://github.com/memcached/memcached/wiki/Proxy

This change revolves around holding context during request execution.
Previously the lua context simply "floated", being carried with a backend IO
request when necessary and resumed during a later callback. It was not
possible to know things for the full duration of a request. It was also not
possible to hold references for the full duration of a request.

For example, previously when lua executes a request against a single pool
(pool(req) syntax), the pool and request objects are held in the coroutine's
execution stack, then the coroutine is suspended. This prevents the pool from
being garbage collected while the request executes. Nothing else explicitly
prevents this from happening.

Nor can we pre-allocate memory and request objects. In a typical request, at
minimum:

- a coroutine thread is allocated
- a request object is allocated
- response objects are allocated
- any strings are allocated
- all of the above must later be garbage collected.

This can add up to a lot of CPU. Object creation is especially slow (setting
metatables) and argument checking can get expensive. This puts a high floor on
the CPU usage of a request. With pre-allocation it should be possible to run a
request from start to finish with zero allocations (and thus zero garbage
collection).

Further, the "async API" available via mcp.await was supposed to be a
temporary workaround before making the API change available in this pull
request. mcp.await is implemented by bolting itself over the existing
system, with hacks to know when to resume the coroutine. Further, it creates
even more allocations:

- the await tracking object
- a table to hold the response objects to be returned to lua

... and itself is limited. There is no way to execute callbacks, to execute
different requests against different pools, to wait on specific pools, and so
on.

Further, there is no way to compose a configuration as a directed graph, as
was done with mcrouter and similar systems. Without this configurations are
less composible, requiring significant code changes to do things like shadow
traffic, pre-warm pools, and so on.

---

We now use a request context to track an incoming request until the response
has been sent to the client. We use a pre-generation stage to allow
pre-allocating objects, caching lookups, creating callback contexts, and so
on. These pre-allocted request contexts and their associated functions are
re-used until the configuration is reloaded and overwritten.

The new API allows for composing configurations as a directed graph,
composing functions that can call either pools or other functions
directly.