core: fix incr/decr/ma failing in some conditions
When an increment or decrement is being applied to a value, we first
parse the string value in the item data into an integer. If this number
is very large (enough to be negative if treated as signed), the
safe_strtoul* functions would then search for a '-' in the string to see
if it was specified as negative.

These functions used strchr() to look for the - sign. The data portion
of an item is _not_ null terminated, meaning the strchr() may wander off
until it finds a '\0' elsewhere in the data. If you're unlucky and it
finds a '-' before a '\0' the arithmetic command may return a
NON_NUMERIC error.

This change caps the length in which to search for the '-' character.
strtoull will itself stop parsing once it reaches a non-numeric digit,
and item data is always terminated by spaces or newlines. So we use the
distance strtoul parsed into the string as the limiter for memchr
instead.

The only reason why it's a scan instead of just looking at the first
character for a '-' is because strtoul* ignores leading whitespace.

Thanks to @feihu-stripe for the report and testing.