proxy: fix `ms` parsing error
In the original string tokenizer the "null" token gets counted as the
final token, so if you want to look for "at least 3 tokens" you need to
check for 4.

In the non-mutative parser used in the internal code the tokens count is
exact, causing a bare "ms foo n\r\n" command to fail.