proxy: locality test + string return fix
we don't process returned raw strings right now. not entirely sure we
should for the first release version of the new API.

thinking about enforcing using a result object if you want to return a
custom value that gets processed properly, but the res object API would
need to expand a bit.