proxy: add global rate limiter
If created during mcp_config_pools() and then passed into
mcp_config_routes(), utilizes a centralized rate limiter.

Generalizes the "proxied reference counting" that the pools system uses
so it can be reused here. This is a system where:

- The object memory is created/managed/owned by the "configuration VM"
- The object is used from worker VM's, thus needs a local lua object to
  represent it.

A "proxy" object is created in each worker VM, which just points back to
the global memory.

When all of the "proxy" objects are GC'ed from all worker VM's, a
management thread is notified and the object is unreferenced from the
configuration VM, allowing it to be GC collected.