proxy: fix unlocked usage of config VM
This bug means under some conditions it's possible for the Lua
configuration VM to be accessed by multiple threads, leading to
corruption. Found by eyeballing it.

In practice this is likely impossible to trigger: only two threads
access the configuration VM:
- Config thread
- Config Manager thread (handles GC'ing of pools)

The manager thread only activates after a configuration reload when old
pool objects are being reaped. It's also only possible to trigger a
configuration reload once per second.

We also _do_ use the configuration lock for most of the manager thread,
the bug just unlocks it early. This means to produce an error you would
need to:

- Issue a reload
- Reload completes
- Some in-flight requests hold the old pool-objects open
- Issue another reload (minimum time 1s)
- A pool object GC's at the same time as the reload is triggered
- The configuration thread serializes behind the manger thread and
  blocks on the config lock during a _very_ narrow window.
- Manager thread releases lock
- Configuration thread acquires lock
- Manager thread issues lua GC calls while config thread is doing other
  things
- Now config VM crash/corruption.

If you're issuing reloads more than a few seconds apart the odds of
triggering this bug are zero.