proxy: fix bug in config reload cleanup phase
In mcp_funcgen_cleanup there is a loop that frees all of the referenced
subrctx's, then a second loop that frees the top level funcgen
references that were held. In the second loop I was treating a subfgen
as a subrctx. Weirdly I'm pretty sure that code had test coverage!

This is obviously the right fix: will figure out what went wrong with
tests later.