slabs: fix CAS changing during page move rescues
When a slab page is being moved between slab classes, it might have
valid memory in it still. In these cases we try to "rescue" the item by
moving it into free memory within the source slab class and updating its
entry in the hash table.

During this move, we were accidentally updating the CAS value. This
one-liner + test fixes this.