tests: add T_MEMD_EXTERNAL
If environment variable T_MEMD_EXTERNAL exists when tests run, a request
to start a memcached daemon will instead print the command line and wait
for the user to start the daemon (ie; in GDB). it will poll and continue
the tests as soon as it can connect.

This is simpler than the old method I was using for T_MEMD_USE_MEMD
where you'd manually start a memcached then run the test.