meta: E flag for overriding CAS value
Ecas means: if operation succeeds, set the item's CAS value to this.
This allows external versioning of item data. This could be a row
version, a time stamp, a crc32 or 64bit data integrity hash, and so on.

overriding the CAS value may make some features not work as well (ie;
runtime enable/disable of `stats sizes`)