core: stop using cas for flush_all
If CAS was enabled, it was used to make `flush_all` more accurate.
Otherwise items set on the same second that a `flush_all` is run would
fail to set.

This restores the old flush method of actively expiring items from the
top of the LRU, but using proper locking. Thus no more use of CAS
comparison and should work properly with our without CAS enabled.