proxy: -o `proxy_arg` to pass cmdline to lua
Assigned to `mcp.start_arg` and available in all threads.

If argument is a normal string, it is assigned as a string.

If the argument contains ':' then it is split into a table. If sections
of the string contain '_' then they are treated as key/value.

`proxy_arg=foo` -> foo.

`proxy_arg=foo_bar:baz' -> { foo = "bar", baz = true }