proxy: rctx:cfd() and `log_req` cfd
log_req() now takes an optional 4th integer argument which is supposed
to be the client file descriptor

log_reqsample() takes an optional 7th integer argument which is supposed
to be the client file descriptor

rctx object now has an `rctx:cfd()` function to return the client fd.

since client objects are lost after a response has been returned
upstream, we have to store the original cfd in the request context.