proxy: `stats proxybe` to show bad backends
stats command intended for backend state information.

so far just outputs "bad_label" [count] for each backend currently
marked bad. The count is the number of backend objects marked bad.

If in "per worker thread mode" (default) if all workers recognize the
backend as down, the count will be the same as the number of worker
threads. If the IO thread is in use it would be 1.

If a backend has normal status, it is not included in the output.

Also fixes a bug with reconnects:

- If a remote server is listening for new connections but not accepting
  them, the backend will fail in readvalidate and never get marked bad.
- Now we only reset the fail counter and remove bad status if a backend
  is fully validated.