proxy: fix race condition leading to hang
If a lua function schedules extra requests after an initial set of
requests, we need to use a workaround to ensure the second set of
requests is actually submitted to the proxy backends.

This workaround was in the wrong place. If connections are being cut and
reconnected between batches of requests it's possible for the workaround
to trigger on a connection object that has since moved to another
thread.

This patch tightens up the workaround to run before the client
connection has a chance to resume.