proxy: fix IO backlog softlock
If an IO thread got a large depth backlog it would only be able to write
1024-ish items at a time, requiring more requests coming in to continue
the flush backlog. This is invisible in normal traffic conditions since
a sudden burst in traffic would still quickly dequeue with normal
traffic levels.

So if you cause a huge backlog and then _stop_ traffic it would never flush
everything and appear to hang.

This does not apply to cases where the backend cannot finish flushing
due to EWOULDBLOCK. That worked fine already.