proxy: fix short writes caused by mcp.internal
Wasn't filling in the 'tosend' portion of a response structure when
using mcp.internal to create a response object.

This field is used as a fastpath check to see if a response has been
completely transmitted to the socket or not. Since it was zero, we would
always consider an mcp.internal() based response as completely sent,
cutting off the data for larger items.

Expands tests for "simply large" and also a large enough to be chunked
item.