proxy: fix backend depth counter
The "request depth" value for a backend could go negative in some cases.
We used a magic value to attempt to speed up the sub-connection selector
but I don't think that's necessary. Without the magic value we don't
need to reset the depth to 0, so it stays balanced.

Added some new asserts so test and bench suites can find this issue more
easily going forward. Survived the full bench suite with asserts
enabled.