proxy: add mcp.backend_depth_limit(int)
If a backend has a queue depth limt over this amount, fast-fail any
further requests.

A global parallel request limit can mean a single slow or overloaded
backend causes the entire proxy to stop working. As a first layer of
defence the depth for a particular backend should be capped.