proxy: support to-be-closed for result objects
if using mcp.internal(r) to fetch keys, but not returning them to the
user, the underlying item references will normally stick around until
they are garbage collected. Memcached is _not_ designed for this:
references must be held for a short and temporary period of time.

If using a res object that you don't intend to send back to the user, it
must be marked with <close>