proxy: fix stupid write flush bug
If we can't write the entire backend request queue within the IOV limit,
we mark where in the request queue we were and try again later if the
socket is still writeable.

Out of paranoia I had marked the position _before_ the next not-flushed
IO, but that is obviously wrong: any _flushed_ IO can be _removed_ from
the stack after processing a READ event from the socket. If we start
again at the position of a flushed IO it might be reclaimed memory.

Somehow this was extremely hard for the bench suite to reproduce, unless
I cut the IOV limit from 1024 to 32.