proxy: router refactor and cmap fallback
Allows specifying a commandmap on a router as well as a map and default
route. Lookup order is:
- Check map, if entry is map, check that
- If none matched, check fallback command map
- If nothing matched, return default route handler if exists

Also removes the router struct from the funcgen object, saving 32 bytes
and letting me extend the size of the router struct without bloating
every allocated fgen. This will later allow inlining the lookup map into
the router struct for a lookup speedup.

Finally, this adds support for "cmap only routers", which is mostly
useful when working with abstractions like routelib.