tls: quarantine openssl headers
To reduce ifdef soup and proliferation of OpenSSL headers, ensure all
openssl-specific calls and defines are only used within tls.c. This
reduces the scope we need to check for compliance with the openssl API.

This change also reduces the amount of ifdef TLS going on.. There is
still a lot left but this is enough change for now.