proxy: ensure resp.elapsed gets set
If a response is returned because of an error clearing a backend queue
the elapsed time never gets set, leaving it as zero. Unless you're using
mcp.await(), which was setting the elapsed time itself.

would be better to wrap return_io_pending() to cut the calls down to 1.
Should do that later.