proxy: [EXPERIMENTAL] mutator framework
See WARNING at top of `proxy_mutator.c`: this is a non-production-ready
experimental interface that needs further work. It's being merged in
order to facilitate easier testing without having losing significant
time to rebasing or fiddling with branches.

THERE IS NO RISK TO DEPLOYMENT if the code is NOT USED. It does not make
any core changes and the only hooks into files outside its own .c are
just enabling the API in lua.