core: fix some potential crash with libevent code
When a log watcher or lru crawler closes early, we re-arm the
connection's event structure, but we may not have first removed the
event.

This seems unnecessary: instead we just do a blind delete when the
connection closes, which libevent seems happy with.

I've also added an event_del to the default, which I don't believe
gets executed.

Do not understand as of this writing how it's possible to get into this
situation, but someone's test is causing it to happen.