proxy: internally track rctx:*_new objects
We need to issue a request or result object cleanup _after_ a request
context has been completed and returned to the slot cache. Else request
or result buffer memory will hang around until the next time the slot is
used.

This requires internally tracking when these objects are created and
directly freeing them. We do this by adding extra rqu's in the existing
rcontext structure, so we can efficiently loop them instead of having to
go through lua.