proxy: simplify lua GC logic
Do not let the speed of the GC bounce up and down as total memory moves,
only increase. If memory decreases during a GC run the speed will stay
constant.

Simplify some of the logic around the functions; removes a few branches
from fixed overhead.

Make "gc timerpoke" even less likely to run: if the main GC is running
_or_ if memory changes at all, don't run the poker. This should avoid a
situation where the poker very slowly frees memory, preventing the main
GC algorithm from knowing when it should actually run next.

Also make the "timerpoke" not update the internal notes on when to run
the GC next. This keeps the ownership of when to start the GC in one
place, also all other accounting can continue to make sense.