proxy: fix issue with waiting on subreqs
If a rctx enqueues N subrctx's, then issues a wait, but all of the
subrctx's fast fail... if you then issue another wait API call the
parent will get resumed recursively before the child is completed.

Meaning the parent won't resume the coroutine if the subrctx's did _not_
match the wait condition (ie: looking for WAIT_GOOD and got WAIT_ANY).

To be honest the issue is so complex but the fix so simple I'm not going
to attempt to describe it past this point.