core: restore automove == 2 mode
I hate how this is implemented, but a better approach would be to make
the in-memory mover algorithm much more aggressive and remove the
in-line reassign. Instead of removing the feature, that is.

The memory algo could be set to "if any class has evictions, move random
slab class to it, and don't delay" or something.