core: IO subsystem refactoring
Aligns the "Background IO system" closer to what the proxy needs to
operate cleanly. Reduces connection structure size ~10% as well.

Mainly this moves the "queue of work" concept from "per connection
structure" to "per event thread". It then changes the concept of a
"suspended connection" to be "suspended responses", which has a counter
per conn. This allows a single suspended response to internally make
many sub IO's and then later resume the original connection without
overcomplicating things.

Most of the time IO's are enqueued to the event thread and then executed
within the context of a single connection executing requests: meaning we
will almost always submit queued for a single connection even with the
new code.

This also fixes rctx:sleep() and unblocks using mcp.internal as a V2 API
backend.