core: allow batch submission to IO subsystem
If the event loop has many clients to work on, allow processing multiple
clients before submitting IO queues to the backend systems.

This is a massive reduction in syscalls when there are few backend
sockets. A test with many active clients went from p50 of 700us to p50
of 100us, both with p99 of 1ms. 500k rps on a 16 core 7950x.

Bounding the batch submission size with a magic value (20) while I learn
more about what workloads this does and doesn't work for. This may be
less great for extstore than the proxy, so need to be conservative for
now. For the proxy any amount of batching is a win, but too much
batching creates a longer tail latency (p99.9 though, p99 seems
unchanged).