proxy: deduplicate protocol processing code
This commit is step one of... a few... for standardizing the code paths
for protocol handling. We started with three:
- mcmc response parsing
- proxy_request.c request parsing
- proto_text.c (original tokenizer)

Both requests and responses can now be tokenized by the same code from
mcmc. This commit series does this in stages, which will result in
removal of duplicated code and fixing many long standing issues.

In this particular commit the duplicated tokenizer is removed from
proxy_request.c, and mcmc functions are now used in various places.