proto: -X disables `stats detail` as well
Since it's possible to gather information about unrelated keys via
`stats detail dump`, it should be included in the --disable-dump list.

Credit to Wallarm / https://www.wallarm.com/ for pointing this out.

Since there is _no multitenancy guarantee_ in memcached disabling of
dumps is just for layered protection. Commits to disable the other
dump commands were treated as new features and not security
vulnerabilities in the past... so we will continue to do so.

Out of an abundance of caution we're putting out a notice about this
regardless.