proxy: centralize ascii_multiget END removal hack
the ascii multiget handler for proxy has to eat the "END\r\n" strings
from each sub-request and then append a final "END\r\n" - this ended up
with hacks deep down in all of the various handlers and needing to pass
flags through io objects and etc.

Instead we tag the top level request context if it's handling an ascii
mlutiget, and a single hack is activated at response time. Took a while
to figure out but worth it.