crawler: speedup and extstore info
Unrolls the sprintf used for `lru_crawler metadump` since it is a hot
path. Increases dump speed with no crawler sleep by 30%.

Also adds the disk page and disk page offset via `ext_page` and
`ext_offset` if the item is on disk.

Credit to @hemal-shah for the idea on dumping extstore information.