proxy: fix `time_*_millis` overflow on 32bit sys
This fixes a time overflow by casting `tv_sec` to `lua_Integer` (which is `long long` unless explicitly configured otherwise) before multiplying it by 1000, which ensures the result doesn't become negative before being assigned to a `lua_Integer` value and returned.

Concretely, this fixes the test failure in `t/proxyunits.t` on systems that have not yet transitioned to `time64_t` (notably, Debian 12 / Bookworm; that transition should be part of 13 / Trixie, which will likely be released later this year).