tests: attempt to make extstore tests less flaky
on very slow systems we may not wait long enough for background threads
to run, and may end up waiting a long time for them to wake up. So we
changed the item_age from 2 seconds to 0 seconds in hopes of preventing
the bg thread from sleeping. We also watch the sysevents log to ensure
the compaction thread is idle for a while after we expect it to run.