tls: fix potential hang when pipelining sets
If TLS is in use and sets are being pipelined together, it's possible
for OpenSSL to read the next command into pending BIO data while we're
reading the payload of a prior set. We then go to sleep expecting epoll
to wake us up if data is still available on the socket.

We now check for SSL_pending bytes before stopping the state machine
loop.

This _may_ not be the only fix required for this state but it will be
the most common one.