proxy: Add farmhash and modulo distribution
The following changes extends the proxy module to support the farmhash
hash function and a simple modulo distribution function.

The modulo function optionally takes a hash_mask parameter to mask the
hash value before applying the modulo operation.

The farmhash function is a wrapper around the farmhash library.

The test case t/proxyhashdist.t demonstrates the use of the farmhash
and modulo distribution functions combined.