tests: fix low TTL test on slow systems
Was giving 20 seconds of TTL for a "low TTL item test" - the test was to
ensure that the TEMP_LRU gets recovered properly on restart, not that an
item of 20 seconds survives. So we can set that TTL to be much higher
to cope with very slow ARM or emulated systems.