proxy: remove buffer memory limit lock
Removes mostly-uncontended lock that was only required for result buffer
accounting when the IO thread is in use. Now buffer memory may be
overused if an IO thread is used for backends, but only by very small
amounts as requests will still begin to fail immediately.

Instead of locking during the network code we update the counter once
the result is returned to the worker thread.

Lock was called a minimum of 4 times and per request and could be dozens
depending on configuration. Not expecting big performance difference
from an uncontested lock but it should help.